requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
altair==5.4.1
annotated-types==0.7.0
anyio==4.6.2.post1
attrs==24.2.0
blinker==1.8.2
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
dataclasses-json==0.6.7
decouple==0.0.7
distlib==0.3.9
distro==1.9.0
filelock==3.16.1
Flask==3.0.3
Flask-Cors==5.0.0
frozenlist==1.5.0
gitdb==4.0.11
GitPython==3.1.43
groq==0.12.0
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
httpx-sse==0.4.0
idna==3.10
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
jiter==0.7.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
langchain==0.3.7
langchain-community==0.3.7
langchain-core==0.3.17
langchain-groq==0.2.1
langchain-text-splitters==0.3.2
langsmith==0.1.142
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.23.1
mdurl==0.1.2
multidict==6.1.0
mypy-extensions==1.0.0
narwhals==1.12.1
numpy==1.26.4
openai==0.27.0
orjson==3.10.11
packaging==24.1
pandas==2.2.3
pillow==10.4.0
pipenv==2024.4.0
platformdirs==4.3.6
plotly==5.24.1
pluggy==1.5.0
propcache==0.2.0
protobuf==5.28.3
pyarrow==18.0.0
pydantic==2.9.2
pydantic-settings==2.6.1
pydantic_core==2.23.4
pydeck==0.9.1
pydub==0.25.1
Pygments==2.18.0
PyPDF2==3.0.1
pytest==8.3.3
pytest-flask==1.3.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.0.1
pytubefix==8.3.0
pytz==2024.2
PyYAML==6.0.2
referencing==0.35.1
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.9.3
rpds-py==0.20.0
semantic-version==2.10.0
setuptools==75.5.0
setuptools-rust==1.10.2
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
SQLAlchemy==2.0.35
streamlit==1.39.0
tenacity==9.0.0
toml==0.10.2
tornado==6.4.1
tqdm==4.66.6
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
uuid==1.30
virtualenv==20.27.1
watchdog==5.0.3
Werkzeug==3.1.3
wheel==0.45.0
whisper==1.1.10
yarl==1.17.1
youtube-transcript-api==0.6.2
yt-dlp==2024.11.4
